feat(metrics): use native request latency histogram (#11251)

+++ src/Orleans.Core/Runtime/CallbackData.cs
@@ -221,7 +221,7 @@ namespace Orleans.Runtime
         {
             if (_applicationRequestInstruments.AppRequestsLatencyEnabled)
             {
-                var elapsedMilliseconds = (long)shared.TimeProvider.GetElapsedTime(_startTimestamp).TotalMilliseconds;
+                var elapsedMilliseconds = shared.TimeProvider.GetElapsedTime(_startTimestamp).TotalMilliseconds;
                 _applicationRequestInstruments.OnAppRequestsEnd(elapsedMilliseconds);
             }
         }
